Feat/scrum 186 에코스톡 계산 로직 수정

+++ src/main/java/org/phoenix/planet/dto/eco_stock/raw/MemberStockInfo.java
@@ -6,7 +6,7 @@ public record MemberStockInfo(
         Long memberId,
         Long ecoStockId,
         Integer currentTotalQuantity,
-        Long currentTotalAmount,
-        Integer point
+        Double currentTotalAmount,
+        Double point
 ) {
 }
